use the os.ErrDeadlineExceeded for stream deadline errors on Go 1.15
