+++ stream.go
@@ -1,7 +1,6 @@
 package quic
 
 import (
-	"net"
 	"sync"
 	"time"
 
@@ -73,14 +72,6 @@ type stream struct {
 }
 
 var _ Stream = &stream{}
-
-type deadlineError struct{}
-
-func (deadlineError) Error() string   { return "deadline exceeded" }
-func (deadlineError) Temporary() bool { return true }
-func (deadlineError) Timeout() bool   { return true }
-
-var errDeadline net.Error = &deadlineError{}
 
 type streamCanceledError struct {
 	error
