+++ stream_deadline_error_go115.go
@@ -1,0 +1,9 @@
+// +build go1.15
+
+package quic
+
+import (
+	"os"
+)
+
+func (deadlineError) Unwrap() error { return os.ErrDeadlineExceeded }
